+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.2.1
+
+- Adds Dropdown component
+- Adds button type in NavLink
+- Adds accentBold type in ListItem
+
 ## v0.2.10
 
 - Add icon guest and selectArrow